fixed disk manager(tokio) dk what happen using claude

+++ src/common/types.rs
@@ -1,2 +1,2 @@
 pub type FrameId = usize;
-pub type PageId = usize;
+pub type PageId = i32;
